% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R
\name{insert_knitrgraphics}
\alias{insert_knitrgraphics}
\title{Insert knitr::include_graphics}
\usage{
insert_knitrgraphics()
}
\description{
Call this function as an addin to insert a code chunk for knitr::include_graphics at the cursor.
}
