% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R
\name{insert_multicolumn}
\alias{insert_multicolumn}
\title{Insert two-column for slidy}
\usage{
insert_multicolumn()
}
\description{
Call this function as an addin to insert multicolumn syntax for slidy slides at the cursor position.
}
